requirements.txt
PyQt6
cassandra-driver
python-dateutil
sortedcontainers
cryptography